refactor: refactored eth_eureka in relayer-lib
