+++ packages/relayer-lib/Cargo.toml
@@ -26,6 +26,7 @@ tracing     = { workspace = true, default-features = true }
 
 tendermint     = { workspace = true, features = ["std"] }
 tendermint-rpc = { workspace = true, features = ["http-client"] }
+tendermint-light-client-verifier = { workspace = true }
 
 ibc-proto-eureka    = { workspace = true }
 ibc-core-host-types = { workspace = true }
